simulator-requirements.txt
docopt~=0.6.2
pyzmq~=19.0.2
pyyaml~=5.3.1
sortedcontainers~=2.2.2
intervaltree~=3.1.0
tqdm~=4.50.2
redis~=3.5.3
procset~=1.0
simanneal~=0.5.0
z3-solver~=4.8.9.0